feat: Add stylus button mapping and key remapping support
- Add configurable stylus button mapping (primary/secondary/tertiary)
  Supports Left, Middle, Right, Back, Forward mouse buttons
- Add key remapping for Volume Up/Down, Back, Power, Scroll Lock, Caps Lock
- Add custom JSON-based key remapping for arbitrary keycodes
- Add support for 8 stylus buttons (was 3)
- Add BUTTON_BACK and BUTTON_FORWARD to InputStub
- Update native code to handle extended stylus buttons
- Add preferences UI for all new features

+++ app/src/main/cpp/lorie/lorie.h
@@ -137,7 +137,7 @@ typedef union {
         uint16_t pressure;
         int8_t tilt_x, tilt_y;
         int16_t orientation;
-        uint8_t buttons, eraser, mouse;
+        uint16_t buttons, eraser, mouse;
     } stylus;
     struct {
         uint8_t t, enable;
